discovery-provider/requirements.txt
# Linting and type support
black==22.3.0
flake8==4.0.1
isort==5.10.1
mypy==0.910
sqlalchemy-stubs==0.4
types-python-dateutil==0.1.4
types-redis==3.5.4
types-requests==2.25.0
types-pytz==2021.1.2

# the only version that supports websockets v10.1 (needed for solana py + anchor py)
web3==6.0.0b1
Flask==1.0.4
# pin itsdangerous and markupsafe https://github.com/pallets/flask/issues/4455
itsdangerous==1.1.0
markupsafe==1.1.1
psycopg2==2.8.5
SQLAlchemy==1.3.0
alembic==1.4.3
celery[redis]==4.3.0
redis==3.2.0
pytest==6.2.5
SQLAlchemy-Utils==0.37.6
chance==0.110
pytest-cov==2.6.0
pytest-dotenv==0.5.2
elasticsearch==8.1.2
flask-cors==3.0.6
gunicorn==20.0.4
jsonschema==4.4.0
flask-restx==0.4.0
hashids==1.2.0
fakeredis==1.4.2
jsonformatter==0.3.0
pytest-postgresql==2.4.1
eventlet==0.28.0
psutil==5.8.0
pytz==2021.1
prometheus-client==0.13.1
click==8.0.4
numpy==1.20.3
datasketch==1.5.7
opentelemetry-api==1.12.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-instrumentation-requests==0.33b0
opentelemetry-instrumentation-logging==0.33b0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
lxml==4.9.1

nats-py==2.2.0

# Solana support
base58==2.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
construct==2.10.67
idna==2.10
pycparser==2.20
PyNaCl==1.4.0
requests==2.25.1
aiohttp==3.8.1
solana==0.23.1
typing-extensions==3.10.0.0
urllib3==1.26.3
websockets==10.1
# Anchor support
anchorpy==0.8.3

# Test
pytest-mock==3.6.1


# Dev reload
watchdog==2.1.7
